lambda now only returns aws lambda, added lambdaUtils instead
