+++ src/aws/lambda.js
@@ -1,16 +1,5 @@
-const { Lambda } = require('./aws');
-const { isOffline } = require('../config');
+const AWS = require('./aws');
 
-const invoke = async (region, functionName, payload) => {
-    if (isOffline()) {
-        return { StatusCode: 200 };
-    }
+const lambda = new AWS.Lambda();
 
-    const lambda = new Lambda({ region });
-    return lambda.invoke({
-        FunctionName: functionName,
-        Payload: JSON.stringify(payload)
-    }).promise();
-};
-
-module.exports = { invoke };
+module.exports = lambda;
